+++ src/db/sysdb_sudo.h
@@ -60,10 +60,10 @@
                                        | SYSDB_SUDO_FILTER_NGRS
 
 errno_t sysdb_sudo_filter_rules_by_time(TALLOC_CTX *mem_ctx,
-                                        size_t in_num_rules,
+                                        uint32_t in_num_rules,
                                         struct sysdb_attrs **in_rules,
                                         time_t now,
-                                        size_t *_num_rules,
+                                        uint32_t *_num_rules,
                                         struct sysdb_attrs ***_rules);
 
 errno_t
